requirements.txt
# Bibliotecas de manipulação e visualização de dados
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.0
seaborn==0.12.2
scipy==1.13.1
scikit-learn==1.3.1

# Bibliotecas para o backend e integração com MongoDB
flask==2.3.2
pymongo==4.10.1
dnspython==2.7.0
python-dotenv==1.0.0

# Bibliotecas opcionais para desempenho e funcionalidades adicionais (remova se não usar)
uvicorn==0.32.0
fastapi==0.115.3
